update cli minimally so it compiles
